requirements.txt
oslo.config>=1.2.0
stevedore>=0.14

# for timeutils
iso8601>=0.1.9

# for jsonutils
six>=1.5.2

# FIXME(markmc): remove this when the drivers no longer
# import eventlet

eventlet>=0.13.0

# used by openstack/common/gettextutils.py
Babel>=1.3

# for the routing notifier
PyYAML>=3.1.0

# rabbit driver is the default
kombu>=2.4.8

# marconiclient for Marconi driver
-f http://tarballs.openstack.org/python-marconiclient/
python-marconiclient
